Moved test to separate file
